feat: move buttons to ui directory
