+++ src/components/molecules/Pair.tsx
@@ -2,12 +2,12 @@ import React, { FC, useState } from 'react';
 import { Stack, Button, IconButton, Tooltip, Badge } from "@mui/material";
 import { AddPair, deletePair, UpdatePair } from "../../api/workflow";
 import { WorkflowFile } from "maxun-core";
-import { ClearButton } from "../atoms/buttons/ClearButton";
+import { ClearButton } from "../ui/buttons/ClearButton";
 import { GenericModal } from "../atoms/GenericModal";
 import { PairEditForm } from "./PairEditForm";
 import { PairDisplayDiv } from "../atoms/PairDisplayDiv";
-import { EditButton } from "../atoms/buttons/EditButton";
-import { BreakpointButton } from "../atoms/buttons/BreakpointButton";
+import { EditButton } from "../ui/buttons/EditButton";
+import { BreakpointButton } from "../ui/buttons/BreakpointButton";
 import VisibilityIcon from '@mui/icons-material/Visibility';
 import styled from "styled-components";
 import { LoadingButton } from "@mui/lab";
