+++ src/components/molecules/PairDetail.tsx
@@ -6,10 +6,10 @@ import TreeView from '@mui/lab/TreeView';
 import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
 import ChevronRightIcon from '@mui/icons-material/ChevronRight';
 import TreeItem from '@mui/lab/TreeItem';
-import { AddButton } from "../atoms/buttons/AddButton";
+import { AddButton } from "../ui/buttons/AddButton";
 import { WarningText } from "../atoms/texts";
 import NotificationImportantIcon from '@mui/icons-material/NotificationImportant';
-import { RemoveButton } from "../atoms/buttons/RemoveButton";
+import { RemoveButton } from "../ui/buttons/RemoveButton";
 import { AddWhereCondModal } from "./AddWhereCondModal";
 import { UpdatePair } from "../../api/workflow";
 import { useSocketStore } from "../../context/socket";
